// templates/homely/src/schemas/Blog.d.ts
export declare const BlogSchema: z.ZodObject<{
    _id: z.ZodOptional<z.ZodType<ObjectId, z.ZodTypeDef, ObjectId>>;
    title: z.ZodString;
    slug: z.ZodString;
    excerpt: z.ZodString;
    content: z.ZodString;
    coverImage: z.ZodString;
    authorId: z.ZodString;
    tags: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    isPublished: z.ZodDefault<z.ZodBoolean>;
    publishedAt: z.ZodDefault<z.ZodNullable<z.ZodDate>>;
    readTime: z.ZodDefault<z.ZodNumber>;
    views: z.ZodDefault<z.ZodNumber>;
    createdAt: z.ZodDefault<z.ZodDate>;
    updatedAt: z.ZodDefault<z.ZodDate>;
}, "strip", z.ZodTypeAny, {
    tags: string[];
    title: string;
    content: string;
    slug: string;
    coverImage: string;
    createdAt: Date;
    readTime: number;
    excerpt: string;
    updatedAt: Date;
    authorId: string;
    isPublished: boolean;
    publishedAt: Date | null;
    views: number;
    _id?: ObjectId | undefined;
}, {
    title: string;
    content: string;
    slug: string;
    coverImage: string;
    excerpt: string;
    authorId: string;
    _id?: ObjectId | undefined;
    tags?: string[] | undefined;
    createdAt?: Date | undefined;
    readTime?: number | undefined;
    updatedAt?: Date | undefined;
    isPublished?: boolean | undefined;
    publishedAt?: Date | null | undefined;
    views?: number | undefined;
}>;
export declare const CreateBlogSchema: z.ZodObject<Omit<{
    _id: z.ZodOptional<z.ZodType<ObjectId, z.ZodTypeDef, ObjectId>>;
    title: z.ZodString;
    slug: z.ZodString;
    excerpt: z.ZodString;
    content: z.ZodString;
    coverImage: z.ZodString;
    authorId: z.ZodString;
    tags: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    isPublished: z.ZodDefault<z.ZodBoolean>;
    publishedAt: z.ZodDefault<z.ZodNullable<z.ZodDate>>;
    readTime: z.ZodDefault<z.ZodNumber>;
    views: z.ZodDefault<z.ZodNumber>;
    createdAt: z.ZodDefault<z.ZodDate>;
    updatedAt: z.ZodDefault<z.ZodDate>;
}, "_id" | "createdAt" | "readTime" | "updatedAt" | "publishedAt">, "strip", z.ZodTypeAny, {
    tags: string[];
    title: string;
    content: string;
    slug: string;
    coverImage: string;
    excerpt: string;
    authorId: string;
    isPublished: boolean;
    views: number;
}, {
    title: string;
    content: string;
    slug: string;
    coverImage: string;
    excerpt: string;
    authorId: string;
    tags?: string[] | undefined;
    isPublished?: boolean | undefined;
    views?: number | undefined;
}>;
export declare const UpdateBlogSchema: z.ZodObject<{
    tags: z.ZodOptional<z.ZodDefault<z.ZodArray<z.ZodString, "many">>>;
    title: z.ZodOptional<z.ZodString>;
    content: z.ZodOptional<z.ZodString>;
    slug: z.ZodOptional<z.ZodString>;
    coverImage: z.ZodOptional<z.ZodString>;
    readTime: z.ZodOptional<z.ZodDefault<z.ZodNumber>>;
    excerpt: z.ZodOptional<z.ZodString>;
    updatedAt: z.ZodOptional<z.ZodDefault<z.ZodDate>>;
    authorId: z.ZodOptional<z.ZodString>;
    isPublished: z.ZodOptional<z.ZodDefault<z.ZodBoolean>>;
    publishedAt: z.ZodOptional<z.ZodDefault<z.ZodNullable<z.ZodDate>>>;
    views: z.ZodOptional<z.ZodDefault<z.ZodNumber>>;
}, "strip", z.ZodTypeAny, {
    tags?: string[] | undefined;
    title?: string | undefined;
    content?: string | undefined;
    slug?: string | undefined;
    coverImage?: string | undefined;
    readTime?: number | undefined;
    excerpt?: string | undefined;
    updatedAt?: Date | undefined;
    authorId?: string | undefined;
    isPublished?: boolean | undefined;
    publishedAt?: Date | null | undefined;
    views?: number | undefined;
}, {
    tags?: string[] | undefined;
    title?: string | undefined;
    content?: string | undefined;
    slug?: string | undefined;
    coverImage?: string | undefined;
    readTime?: number | undefined;
    excerpt?: string | undefined;
    updatedAt?: Date | undefined;
    authorId?: string | undefined;
    isPublished?: boolean | undefined;
    publishedAt?: Date | null | undefined;
    views?: number | undefined;
}>;
export declare const BlogFiltersSchema: z.ZodObject<{
    isPublished: z.ZodOptional<z.ZodBoolean>;
    tags: z.ZodOptional<z.ZodArray<z.ZodString, "many">>;
    limit: z.ZodDefault<z.ZodNumber>;
    skip: z.ZodDefault<z.ZodNumber>;
}, "strip", z.ZodTypeAny, {
    limit: number;
    skip: number;
    tags?: string[] | undefined;
    isPublished?: boolean | undefined;
}, {
    tags?: string[] | undefined;
    limit?: number | undefined;
    isPublished?: boolean | undefined;
    skip?: number | undefined;
}>;
export type Blog = z.infer<typeof BlogSchema>;
export type CreateBlogInput = z.infer<typeof CreateBlogSchema>;
export type UpdateBlogInput = z.infer<typeof UpdateBlogSchema>;
export type BlogFilters = z.infer<typeof BlogFiltersSchema>;